move MockedAudioData class to test helpers module
